chore: Split global.d.ts to meke it more stable
